# apps/desktop/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ABM Invoice Analytics Engine",
  "description": "Local engine API for the ABM Invoice Analytics desktop app. Loopback only.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/setup/admin": {
   "post": {
    "tags": [
     "system"
    ],
    "summary": "Setup Admin",
    "operationId": "setup_admin_setup_admin_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SetupAdmin"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Setup Admin Setup Admin Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Login"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Login Auth Login Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/ingest/files": {
   "post": {
    "tags": [
     "ingest"
    ],
    "summary": "Ingest Files",
    "operationId": "ingest_files_ingest_files_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_ingest_files_ingest_files_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Ingest Files Ingest Files Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/ingest/jobs/{job_id}": {
   "get": {
    "tags": [
     "ingest"
    ],
    "summary": "Ingest Job",
    "operationId": "ingest_job_ingest_jobs__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Ingest Job Ingest Jobs  Job Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/detect/sweep/{job_id}": {
   "get": {
    "tags": [
     "detect"
    ],
    "summary": "Sweep Status",
    "operationId": "sweep_status_detect_sweep__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Sweep Status Detect Sweep  Job Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_ingest_files_ingest_files_post": {
    "properties": {
     "files": {
      "items": {
       "type": "string",
       "contentMediaType": "application/octet-stream"
      },
      "type": "array",
      "title": "Files"
     },
     "options": {
      "type": "string",
      "title": "Options",
      "default": "{}"
     },
     "template_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template Id"
     },
     "mapping": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mapping"
     }
    },
    "type": "object",
    "required": [
     "files"
    ],
    "title": "Body_ingest_files_ingest_files_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Login": {
    "properties": {
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "Login"
   },
...
   "SetupAdmin": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "title": "Username"
     },
     "display_name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Display Name"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 10,
      "title": "Password"
     },
     "create_recovery_key": {
      "type": "boolean",
      "title": "Create Recovery Key",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "username",
     "display_name",
     "password"
    ],
    "title": "SetupAdmin"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}